Fixin the post install and uninstall scripts

+++ Arista/post_uninstall_config.sh
@@ -8,8 +8,10 @@ if [[ "$SYSTEMD_INUSE" != "$TRUE" ]]; then
 	exit 0
 fi
 
-systemctl stop telegraf
-systemctl disable telegraf
-TELEGRAF_SERVICE_OVERRIDE=/usr/lib/systemd/system/telegraf.service
-[[ -f ${TELEGRAF_SERVICE_OVERRIDE} ]] && rm -f ${TELEGRAF_SERVICE_OVERRIDE}
-[[ -f ${DHCLIENT_HOOKS_FILE} ]] && sed -i '/start telegraf/d' ${DHCLIENT_HOOKS_FILE}
+# Clean up only when rpm is uninstalled not during upgrade
+if [[ "$1" == "0" ]]; then 
+    systemctl stop telegraf
+    systemctl disable telegraf
+    TELEGRAF_SERVICE_OVERRIDE=/usr/lib/systemd/system/telegraf.service
+    [[ -f ${DHCLIENT_HOOKS_FILE} ]] && sed -i '/start telegraf/d' ${DHCLIENT_HOOKS_FILE}
+fi
